[refactor] refactor Package and file names
