+++ src/main/java/Opensource_SW_Project/Project/service/MemberService/MemberQueryService.java
@@ -1,4 +1,4 @@
-package Opensource_SW_Project.Project.service.UserService;
+package Opensource_SW_Project.Project.service.MemberService;
 
 import Opensource_SW_Project.Project.domain.Member;
 import Opensource_SW_Project.Project.web.dto.Member.MemberRequestDTO;
